requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
PyMySQL==0.9.3
Simplejson==3.17.0
Jinja2==2.11.1
python-dotenv==0.13.0
Werkzeug==1.0.1
DBUtils==1.3
WTForms-Components==0.10.4
pandas==1.0.3
python-dotenv==0.13.0
flask-login==0.5.0
cryptography==2.9.2
email_validator==1.1.1
pyOpenSSL==19.1.0
oauthlib==3.1.0
requests==2.24.0
